(refactor) Map routeResponse to populate info windows
Also factored out all the utility functions used in the InfoWindow
component into a separate utility file
