+++ client/src/components/InfoWindow/InfoWindow.jsx
@@ -1,61 +1,8 @@
 import React, { Component } from 'react';
 import deepEqual from 'deep-equal';
 import DestinationIcon from 'material-ui/svg-icons/communication/location-on';
+import infoWindowHelpers from '../utils/infoWindowHelpers';
 import './InfoWindow.css';
-
-const getDurationDifference = (currentRouteDuration, alternateRouteDuration) => {
-  let durationDifference;
-  const difference = currentRouteDuration - alternateRouteDuration;
-  const description = difference < 0 ? 'slower' : 'faster';
-  const absoluteDifference = Math.abs(difference);
-
-  if (absoluteDifference < 60) {
-    durationDifference = `${absoluteDifference} sec${(absoluteDifference !== 1 ? 's' : '')} ${description}`;
-  } else if (absoluteDifference < 60 * 60) {
-    const minutes = Math.floor(absoluteDifference / 60);
-    durationDifference = `${minutes}min ${description}`;
-  }
-
-  return durationDifference;
-};
-
-const getRiskDescription = (risk) => {
-  const absoluteRisk = Math.abs(risk);
-  const smooth = risk > 5 && risk <= 15 ? 'smoother' : 'smooth';
-  const description = risk > 0 ? 'choppy' : `${smooth}`;
-
-  switch (risk) {
-    case absoluteRisk <= 5:
-      return '<i>Same swell...</i>';
-
-    case absoluteRisk > 5 && absoluteRisk <= 15:
-      return `<i>A little ${description}</i> `;
-
-    case absoluteRisk > 15 && absoluteRisk <= 25:
-      return `<i>Very ${description}</i>`;
-
-    default:
-      return `<i>Hella ${description}</i>`;
-  }
-};
-
-const displayRiskDifference = (currentRouteAvgRisk, alternateRouteAvgRisk) => {
-  // Risk description will be only displayed when the other route is selected so these should be in relation
-  // to the other route
-  const normalizedRiskDifference = Math.floor(((currentRouteAvgRisk - alternateRouteAvgRisk) / alternateRouteAvgRisk) * 100);
-
-  return getRiskDescription(normalizedRiskDifference);
-};
-
-const displayHoursMinutes = (seconds) => {
-  const totalMinutes = Math.floor(seconds / 60);
-  const hours = Math.floor(totalMinutes / 60);
-  const minutesLeft = totalMinutes % 60;
-  const hoursDescription = hours > 10 ? 's' : '';
-  return `<span class='info-window-time'><b>${hours}hr${hoursDescription} ${minutesLeft}min</b></span>`;
-};
-
-const displayMiles = meters => (Math.round(meters * 0.000621371));
 
 class InfoWindow extends Component {
   constructor(props) {
@@ -76,7 +23,7 @@ class InfoWindow extends Component {
       this.renderInfoWindow();
     } else {
       // If route is currently not displayed, show time travel difference
-      if (this.props.routeIndex !== this.props.currentRouteDisplayedIndex) {
+      if (this.props.routeIndex !== this.props.currentRouteDisplayedIndex && this.props.alternateRoute) {
         this.state.infoWindow.setContent(`<div class='info-window-not-selected'><b>${this.state.durationDifference}</b></br>${this.state.riskAssessment}</div>`);
         this.state.infoWindow.setZIndex(-10);
       } else {
@@ -109,24 +56,31 @@ class InfoWindow extends Component {
       const routeIsCurrentlyDisplayed = this.props.routeIndex ===
                                         this.props.currentRouteDisplayedIndex;
 
+      // If route currently displayed, anchor the info window one-third into the route
       let infoWindowAnchor;
       if (routeIsCurrentlyDisplayed) {
-        // If route currently displayed, anchor the info window one-third into the route
         infoWindowAnchor = Math.floor(this.props.route.path.length / 3);
       } else {
-        // If route not currently displayed, anchor the info window two-thirds into the route
+        // If route is not currently displayed, anchor the info window two-thirds into the route
         infoWindowAnchor = Math.floor((this.props.route.path.length / 3) * 2);
       }
 
       const lat = this.props.route.path[infoWindowAnchor][0];
       const lng = this.props.route.path[infoWindowAnchor][1];
 
-      const duration = displayHoursMinutes(this.props.route.duration);
-      const distance = displayMiles(this.props.route.distance);
-      const differenceDisplay = getDurationDifference(this.props.route.duration,
-                                                      this.props.alternateRoute.duration);
-      const riskAssessment = displayRiskDifference(this.props.route.averageRisk,
-                                                   this.props.alternateRoute.averageRisk);
+      const duration = infoWindowHelpers.displayHoursMinutes(this.props.route.duration);
+      const distance = infoWindowHelpers.displayMiles(this.props.route.distance);
+
+      
+      // Check for duration and risk difference if an alternate route exists
+      let differenceDisplay;
+      let riskAssessment;
+      if (this.props.alternateRoute) {
+        differenceDisplay = infoWindowHelpers.getDurationDifference(this.props.route.duration,
+                                                        this.props.alternateRoute.duration);
+        riskAssessment = infoWindowHelpers.displayRiskDifference(this.props.route.averageRisk,
+                                                     this.props.alternateRoute.averageRisk);  
+      }
 
       const contentClass = routeIsCurrentlyDisplayed ? 'info-window-selected' : 'info-window-not-selected';
 
@@ -144,9 +98,9 @@ class InfoWindow extends Component {
       this.setState({
         infoWindow,
         duration,
-        durationDifference: differenceDisplay,
+        durationDifference: differenceDisplay || null,
         distance,
-        riskAssessment,
+        riskAssessment: riskAssessment || null,
       });
     }
   }
@@ -176,7 +130,7 @@ InfoWindow.propTypes = {
     duration: React.PropTypes.number.isRequired,
     googleMapsUrl: React.PropTypes.string.isRequired,
     path: React.PropTypes.array.isRequired,
-  }).isRequired,
+  }),
   currentRouteDisplayedIndex: React.PropTypes.number.isRequired,
   routeIndex: React.PropTypes.number.isRequired,
 };
